split searchSection out into separate local component
